+++ walkers/src/lib.rs
@@ -8,7 +8,7 @@ pub mod providers;
 mod tiles;
 mod zoom;
 
-pub use map::{Center, Map, MapMemory, Projector};
+pub use map::{Center, Map, MapMemory, Plugin, Projector};
 pub use mercator::{screen_to_position, Position, PositionExt};
 pub use tiles::Tiles;
 pub use zoom::Zoom;