+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
  * Tile sources are now defined via `TileSource` trait, instead of `Fn`.
+ * New `Plugin` trait and `Map::with_plugin` function, which replaces `Map::with_drawer`.
 
 ## 0.8.0